refactor: Convert OrganisationsSwitch into a resource component and update its use
