+++ app/components/dashboard/layout/mobile-nav.tsx
@@ -9,12 +9,12 @@ import { CaretRight as CaretRightIcon } from '@phosphor-icons/react/dist/ssr/Car
 import * as React from 'react'
 
 import { icons } from './nav-icons'
-import { OrganisationsSwitch } from './organisations-switch'
 import { RouterLink } from '@/components/core/link'
 import { Logo } from '@/components/core/logo'
 import { usePathname } from '@/hooks/use-pathname'
 import { isNavItemActive } from '@/lib/is-nav-item-active'
 import { paths } from '@/paths'
+import { OrganisationsSwitch } from '@/routes/resources+/organisations-switch'
 import { type NavItemConfig } from '@/types/nav'
 
 export interface MobileNavProps {
